// src/types/basic_manage/contact.d.ts
// 联系人属性
export interface ContactItemProps {
    
    // 联系人ID
    ContactId:string;
    // 联系人
    ContactName:string;
    // 职能
    JobFunction:string;
    // 航线
    Route:string;
    // 电话
    Phone:string;
    // 地址
    Address:string;
    // 邮件地址
    Email:string;
    // 抄送地址
    CcEmail:string;
    // 手机
    Mobile:string;
    // 推送节点
    PushNode:string;
    // 推送方式
    PushMethod:string;
    // 备注
    Remarks:string;
    // QQ
    QqNumber:string;
    // 微信号
    WechatId:string;
    // Skype
    SkypeId:string;
}